requirement.txt
argparse
requests
websocket-client
pygments
